+++ alpha/apps/kaltura/lib/batch2/bcdl/kBusinessConvertDL.php
@@ -722,4 +722,5 @@ class kBusinessConvertDL
 		}
 		return $cmdLine;
 	}
+
 }