+++ deployment/updates/scripts/add_permissions/2018_03_12_allow_media_server_to_get_conversionprofile.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * @package deployment
+ * @subpackage falcon.roles_and_permissions
+ * 
+ * Add permissions to caption asset
+ */
+
+$script = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$config = realpath(dirname(__FILE__)) . '/../../../permissions/service.conversionprofile.ini';
+passthru("php $script $config");